update staging giftara

+++ wp-content/plugins/all-in-one-wp-security-and-firewall/classes/wp-security-user-login.php
@@ -577,7 +577,7 @@ class AIOWPSecurity_User_Login {
 				$logout_time_interval_val_seconds = $logout_time_interval_value * 60;
 				if ($diff > $logout_time_interval_val_seconds) {
 					$aio_wp_security->debug_logger->log_debug("Force Logout - This user logged in more than (".$logout_time_interval_value.") minutes ago. Doing a force log out for the user with username: ".$current_user->user_login);
-					$this->wp_logout_action_handler($user_id); //this will register the logout time/date in the logout_date column
+					$this->wp_logout_action_handler($user_id, true); //this will register the logout time/date in the logout_date column
 
 
 					$curr_page_url = AIOWPSecurity_Utility::get_current_page_url();
@@ -691,7 +691,9 @@ class AIOWPSecurity_User_Login {
 			return;
 		}
 
-		$this->delete_logged_in_user($user->ID);
+		if (empty(get_user_meta($user_id, 'session_tokens', true))) {
+			$this->delete_logged_in_user($user->ID);
+		}
 
 		if (is_super_admin($user->ID)) {
 			$logging_out_of_correct_site = true;
@@ -772,7 +774,7 @@ class AIOWPSecurity_User_Login {
 		$unlock_secret_string = $aio_wp_security->configs->get_value('aiowps_unlock_request_secret_key');
 		$current_time = time();
 		$enc_result = base64_encode($current_time.$unlock_secret_string);
-		$unlock_request_form .= '<form method="post" action=""><div style="padding-bottom:10px;"><input type="hidden" name="aiowps-unlock-string-info" id="aiowps-unlock-string-info" value="'.$enc_result.'" />';
+		$unlock_request_form .= '<form method="post" action="">'.wp_nonce_field('aios-unlock-nonce', '_wpnonce', true, false).'<div style="padding-bottom:10px;"><input type="hidden" name="aiowps-unlock-string-info" id="aiowps-unlock-string-info" value="'.$enc_result.'" />';
 		$unlock_request_form .= '<input type="hidden" name="aiowps-unlock-temp-string" id="aiowps-unlock-temp-string" value="'.$current_time.'" />';
 		// phpcs:ignore WordPress.Security.NonceVerification.Missing -- No nonce.
 		if (isset($_POST['woocommerce-login-nonce'])) {
@@ -936,7 +938,7 @@ class AIOWPSecurity_User_Login {
 		if (empty($auth_cookie)) return; //check if auth cookie is empty, meaning login was not successful
 		$expiration = $expire > 0 ? $expire : $expiration;
 
-		if (is_multisite() && !is_super_admin()) {
+		if (is_multisite() && !is_super_admin($user_id)) {
 			$user_blog = get_active_blog_for_user($user_id);
 			switch_to_blog($user_blog->blog_id); // switch to user blog incase they try to log in from wrong subsite
